nlp_model/requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
pillow>=10.0.0
pytesseract>=0.3.10

# NLP & Text Processing
nltk>=3.8.1
spacy>=3.7.0
textblob>=0.17.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
